fix(dashboard): move old lhpath tests to variables test
